backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-openai>=1.1.11",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
